Separated DOM code, added SAX code
Added XMLParserSAX (and its required DMOZHandler) - a basic
implementation of SAX APIm which does nothing yet but parse the DMOZ
data (Now successfully. Yet more useless data had to be removed by
hand).

It's unclear whether we'll be needing the DOM tree in the future, so the
old XMLParser has not been removed but instead renamed to XMLParserDOM.